Move signer 0 test [skip ci]
