+++ types/spectest/all_tests.go
@@ -24,6 +24,8 @@ var AllTests = []SpecTest{
 
 	partialsigmessage.Encoding(),
 	partialsigmessage.InvalidMsg(),
+	partialsigmessage.MessageSigner0(),
+	partialsigmessage.SignedMsgSigner0(),
 
 	share.Encoding(),
 	//share.HasQuorum(),
