+++ types/spectest/tests/partialsigmessage/msg_signer_zero.go
@@ -1,4 +1,4 @@
-package messages
+package partialsigmessage
 
 import (
 	"github.com/attestantio/go-eth2-client/spec"
