--- types/spectest/tests/partialsigmessage/partial_sig_signer_zero.go
@@ -1,6 +1,0 @@
-package partialsigmessage
-
-// PartialSigSignerZero tests signer ID 0
-func PartialSigSignerZero() *MsgSpecTest {
-	panic("implement")
-}
