+++ types/spectest/tests/partialsigmessage/signed_msg_signer_zero.go
@@ -1,4 +1,4 @@
-package messages
+package partialsigmessage
 
 import (
 	"github.com/attestantio/go-eth2-client/spec"
